refactor: move storage/ to internal/storage/
