+++ internal/ui/ui.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"dndgoldtracker/internal/party"
-	"dndgoldtracker/storage"
+	"dndgoldtracker/internal/storage"
 
 	"charm.land/bubbles/v2/table"
 	"charm.land/bubbles/v2/textinput"
